[FIX] format: move SET_DECIMAL out of core
Handling "SET_DECIMAL" command depends on the evaluated value of cells.

The values are not guaranteed to be the same for every users because
they can depends on: asynchronous network calls, user access rights.

Handling the same SET_DECIMAL command can therefore lead to different
results for different users which would be de-synchronized.

We have a task planned to completely remove the evaluation from core to
prevent such mistakes in the future. See 2813749
