+++ src/plugins/ui/format.ts
@@ -1,0 +1,168 @@
+import { DATETIME_FORMAT } from "../../constants";
+import { maximumDecimalPlaces } from "../../helpers";
+import { Mode } from "../../model";
+import { CellValueType, Command, UID, Zone } from "../../types/index";
+import { UIPlugin } from "../ui_plugin";
+
+export class FormatPlugin extends UIPlugin {
+  static modes: Mode[] = ["normal"];
+  // ---------------------------------------------------------------------------
+  // Command Handling
+  // ---------------------------------------------------------------------------
+
+  handle(cmd: Command) {
+    switch (cmd.type) {
+      case "SET_DECIMAL":
+        this.setDecimal(cmd.sheetId, cmd.target, cmd.step);
+        break;
+    }
+  }
+
+  /**
+   * This function allows to adjust the quantity of decimal places after a decimal
+   * point on cells containing number value. It does this by changing the cells
+   * format. Values aren't modified.
+   *
+   * The change of the decimal quantity is done one by one, the sign of the step
+   * variable indicates whether we are increasing or decreasing.
+   *
+   * If several cells are in the zone, the format resulting from the change of the
+   * first cell (with number type) will be applied to the whole zone.
+   */
+  private setDecimal(sheetId: UID, zones: Zone[], step: number) {
+    // Find the first cell with a number value and get the format
+    const numberFormat = this.searchNumberFormat(sheetId, zones);
+    if (numberFormat !== undefined) {
+      // Depending on the step sign, increase or decrease the decimal representation
+      // of the format
+      const newFormat = this.changeDecimalFormat(numberFormat, step);
+      // Apply the new format on the whole zone
+      this.dispatch("SET_FORMATTING", {
+        sheetId,
+        target: zones,
+        format: newFormat,
+      });
+    }
+  }
+
+  /**
+   * Take a range of cells and return the format of the first cell containing a
+   * number value. Returns a default format if the cell hasn't format. Returns
+   * undefined if no number value in the range.
+   */
+  private searchNumberFormat(sheetId: UID, zones: Zone[]): string | undefined {
+    for (let zone of zones) {
+      for (let row = zone.top; row <= zone.bottom; row++) {
+        for (let col = zone.left; col <= zone.right; col++) {
+          const cell = this.getters.getCell(sheetId, col, row);
+          if (
+            cell?.evaluated.type === CellValueType.number &&
+            !cell.format?.match(DATETIME_FORMAT) // reject dates
+          ) {
+            return cell.format || this.setDefaultNumberFormat(cell.evaluated.value);
+          }
+        }
+      }
+    }
+    return undefined;
+  }
+
+  /**
+   * Function used to give the default format of a cell with a number for value.
+   * It is considered that the default format of a number is 0 followed by as many
+   * 0 as there are decimal places.
+   *
+   * Example:
+   * - 1 --> '0'
+   * - 123 --> '0'
+   * - 12345 --> '0'
+   * - 42.1 --> '0.0'
+   * - 456.0001 --> '0.0000'
+   */
+  private setDefaultNumberFormat(cellValue: number): string {
+    const strValue = cellValue.toString();
+    const parts = strValue.split(".");
+    if (parts.length === 1) {
+      return "0";
+    }
+    return "0." + Array(parts[1].length + 1).join("0");
+  }
+
+  /**
+   * This function take a cell format representation and return a new format representation
+   * with more or less decimal places.
+   *
+   * If the format doesn't look like a digital format (means that not contain '0')
+   * or if this one cannot be increased or decreased, the returned format will be
+   * the same.
+   *
+   * This function aims to work with all possible formats as well as custom formats.
+   *
+   * Examples of format changed by this function:
+   * - "0" (step = 1) --> "0.0"
+   * - "0.000%" (step = 1) --> "0.0000%"
+   * - "0.00" (step = -1) --> "0.0"
+   * - "0%" (step = -1) --> "0%"
+   * - "#,##0.0" (step = -1) --> "#,##0"
+   * - "#,##0;0.0%;0.000" (step = 1) --> "#,##0.0;0.00%;0.0000"
+   */
+  private changeDecimalFormat(format: string, step: number): string {
+    const sign = Math.sign(step);
+    // According to the representation of the cell format. A format can contain
+    // up to 4 sub-formats which can be applied depending on the value of the cell
+    // (among positive / negative / zero / text), each of these sub-format is separated
+    // by ';' in the format. We need to make the change on each sub-format.
+    const subFormats = format.split(";");
+    let newSubFormats: string[] = [];
+
+    for (let subFormat of subFormats) {
+      const decimalPointPosition = subFormat.indexOf(".");
+      const exponentPosition = subFormat.toUpperCase().indexOf("E");
+      let newSubFormat: string;
+
+      // the 1st step is to find the part of the zeros located before the
+      // exponent (when existed)
+      const subPart = exponentPosition > -1 ? subFormat.slice(0, exponentPosition) : subFormat;
+      const zerosAfterDecimal =
+        decimalPointPosition > -1 ? subPart.slice(decimalPointPosition).match(/0/g)!.length : 0;
+
+      // the 2nd step is to add (or remove) zero after the last zeros obtained in
+      // step 1
+      const lastZeroPosition = subPart.lastIndexOf("0");
+      if (lastZeroPosition > -1) {
+        if (sign > 0) {
+          // in this case we want to add decimal information
+          if (zerosAfterDecimal < maximumDecimalPlaces) {
+            newSubFormat =
+              subFormat.slice(0, lastZeroPosition + 1) +
+              (zerosAfterDecimal === 0 ? ".0" : "0") +
+              subFormat.slice(lastZeroPosition + 1);
+          } else {
+            newSubFormat = subFormat;
+          }
+        } else {
+          // in this case we want to remove decimal information
+          if (zerosAfterDecimal > 0) {
+            // remove last zero
+            newSubFormat =
+              subFormat.slice(0, lastZeroPosition) + subFormat.slice(lastZeroPosition + 1);
+            // if a zero always exist after decimal point else remove decimal point
+            if (zerosAfterDecimal === 1) {
+              newSubFormat =
+                newSubFormat.slice(0, decimalPointPosition) +
+                newSubFormat.slice(decimalPointPosition + 1);
+            }
+          } else {
+            // zero after decimal isn't present, we can't remove zero
+            newSubFormat = subFormat;
+          }
+        }
+      } else {
+        // no zeros are present in this format, we do nothing
+        newSubFormat = subFormat;
+      }
+      newSubFormats.push(newSubFormat);
+    }
+    return newSubFormats.join(";");
+  }
+}
